+++ Editor/Drawers/FindComponentAttributeDrawer.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Reflection;
 using SaintsField.Editor.Core;
 using SaintsField.Editor.Utils;
@@ -25,7 +27,7 @@ namespace SaintsField.Editor.Drawers
             int index,
             OnGUIPayload onGUIPayload, FieldInfo info, object parent)
         {
-            (string error, Object result) = DoCheckComponent(property, saintsAttribute, info);
+            (string error, Object result) = DoCheckComponent(property, saintsAttribute, info, parent);
             if (error != "")
             {
                 _error = error;
@@ -48,17 +50,52 @@ namespace SaintsField.Editor.Drawers
             ISaintsAttribute saintsAttribute, FieldInfo info, object parent) => _error == ""? position: ImGuiHelpBox.Draw(position, _error, EMessageType.Error);
         #endregion
 
-        private static (string error, Object result) DoCheckComponent(SerializedProperty property, ISaintsAttribute saintsAttribute, FieldInfo info)
+        private static (string error, Object result) DoCheckComponent(SerializedProperty property, ISaintsAttribute saintsAttribute, FieldInfo info, object parent)
         {
-            if (property.objectReferenceValue != null)
+            SerializedProperty targetProperty = property;
+            Type fieldType = info.FieldType;
+            Type interfaceType = null;
+            if (property.propertyType == SerializedPropertyType.Generic)
+            {
+                object propertyValue = SerializedUtils.GetValue(property, info, parent);
+
+                if (propertyValue is IWrapProp wrapProp)
+                {
+                    Type mostBaseType = SaintsInterfaceDrawer.GetMostBaseType(wrapProp.GetType());
+                    if (mostBaseType.IsGenericType && mostBaseType.GetGenericTypeDefinition() == typeof(SaintsInterface<,>))
+                    {
+                        IReadOnlyList<Type> genericArguments = mostBaseType.GetGenericArguments();
+                        if (genericArguments.Count == 2)
+                        {
+                            interfaceType = genericArguments[1];
+                        }
+                    }
+                    targetProperty = property.FindPropertyRelative(wrapProp.EditorPropertyName) ??
+                                     SerializedUtils.FindPropertyByAutoPropertyName(property,
+                                         wrapProp.EditorPropertyName);
+
+                    if(targetProperty == null)
+                    {
+                        return ($"{wrapProp.EditorPropertyName} not found in {property.propertyPath}", null);
+                    }
+
+                    SerializedUtils.FieldOrProp wrapFieldOrProp = Util.GetWrapProp(wrapProp);
+                    fieldType = wrapFieldOrProp.IsField
+                        ? wrapFieldOrProp.FieldInfo.FieldType
+                        : wrapFieldOrProp.PropertyInfo.PropertyType;
+                    if (interfaceType != null && fieldType != typeof(Component) && !fieldType.IsSubclassOf(typeof(Component)) && typeof(Component).IsSubclassOf(fieldType))
+                    {
+                        fieldType = typeof(Component);
+                    }
+                }
+            }
+
+            if (targetProperty.objectReferenceValue != null)
             {
                 return ("", null);
             }
 
             FindComponentAttribute findComponentAttribute = (FindComponentAttribute) saintsAttribute;
-            Type fieldType = info.FieldType;
-
-            // Type type = findComponentAttribute.CompType ?? fieldType;
 
             Transform transform;
             switch (property.serializedObject.targetObject)
@@ -88,13 +125,10 @@ namespace SaintsField.Editor.Drawers
                     break;
                 }
 
-                if (fieldType == typeof(Transform))
-                {
-                    componentInChildren = findTarget;
-                    break;
-                }
+                componentInChildren = interfaceType == null
+                    ? findTarget.GetComponent(fieldType)
+                    : findTarget.GetComponents(fieldType).FirstOrDefault(interfaceType.IsInstanceOfType);
 
-                componentInChildren = findTarget.GetComponent(fieldType);
                 if (componentInChildren != null)
                 {
                     break;
@@ -106,7 +140,7 @@ namespace SaintsField.Editor.Drawers
                 return ($"No {fieldType} found in paths: {string.Join(", ", findComponentAttribute.Paths)}", null);
             }
 
-            property.objectReferenceValue = componentInChildren;
+            targetProperty.objectReferenceValue = componentInChildren;
 
             return ("", componentInChildren);
         }
@@ -140,7 +174,7 @@ namespace SaintsField.Editor.Drawers
 #if SAINTSFIELD_DEBUG && SAINTSFIELD_DEBUG_DRAW_PROCESS_FIND_COMPONENT
             Debug.Log($"FindComponent DrawPostFieldUIToolkit for {property.propertyPath}");
 #endif
-            (string error, Object result) = DoCheckComponent(property, saintsAttribute, info);
+            (string error, Object result) = DoCheckComponent(property, saintsAttribute, info, parent);
             HelpBox helpBox = container.Q<HelpBox>(NamePlaceholder(property, index));
             if (error != helpBox.text)
             {