+++ Editor/Drawers/GetComponentInParentsAttributeDrawer.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Reflection;
 using SaintsField.Editor.Core;
 using SaintsField.Editor.Utils;
@@ -25,7 +27,7 @@ namespace SaintsField.Editor.Drawers
             int index,
             OnGUIPayload onGUIPayload, FieldInfo info, object parent)
         {
-            (string error, UnityEngine.Object result) = DoCheckComponent(property, (GetComponentInParentsAttribute)saintsAttribute, info);
+            (string error, UnityEngine.Object result) = DoCheckComponent(property, (GetComponentInParentsAttribute)saintsAttribute, info, parent);
             if (error != "")
             {
                 _error = error;
@@ -47,15 +49,51 @@ namespace SaintsField.Editor.Drawers
             ISaintsAttribute saintsAttribute, FieldInfo info, object parent) => _error == ""? position: ImGuiHelpBox.Draw(position, _error, EMessageType.Error);
         #endregion
 
-        private static (string error, UnityEngine.Object result) DoCheckComponent(SerializedProperty property, GetComponentInParentsAttribute getComponentInParentsAttribute, FieldInfo info)
+        private static (string error, UnityEngine.Object result) DoCheckComponent(SerializedProperty property, GetComponentInParentsAttribute getComponentInParentsAttribute, FieldInfo info, object parent)
         {
-            if (property.objectReferenceValue != null)
+            SerializedProperty targetProperty = property;
+            Type fieldType = info.FieldType;
+            Type interfaceType = null;
+            if (property.propertyType == SerializedPropertyType.Generic)
+            {
+                object propertyValue = SerializedUtils.GetValue(property, info, parent);
+
+                if (propertyValue is IWrapProp wrapProp)
+                {
+                    Type mostBaseType = SaintsInterfaceDrawer.GetMostBaseType(wrapProp.GetType());
+                    if (mostBaseType.IsGenericType && mostBaseType.GetGenericTypeDefinition() == typeof(SaintsInterface<,>))
+                    {
+                        IReadOnlyList<Type> genericArguments = mostBaseType.GetGenericArguments();
+                        if (genericArguments.Count == 2)
+                        {
+                            interfaceType = genericArguments[1];
+                        }
+                    }
+                    targetProperty = property.FindPropertyRelative(wrapProp.EditorPropertyName) ??
+                                     SerializedUtils.FindPropertyByAutoPropertyName(property,
+                                         wrapProp.EditorPropertyName);
+
+                    if(targetProperty == null)
+                    {
+                        return ($"{wrapProp.EditorPropertyName} not found in {property.propertyPath}", null);
+                    }
+
+                    SerializedUtils.FieldOrProp wrapFieldOrProp = Util.GetWrapProp(wrapProp);
+                    fieldType = wrapFieldOrProp.IsField
+                        ? wrapFieldOrProp.FieldInfo.FieldType
+                        : wrapFieldOrProp.PropertyInfo.PropertyType;
+
+                    if (interfaceType != null && fieldType != typeof(Component) && !fieldType.IsSubclassOf(typeof(Component)) && typeof(Component).IsSubclassOf(fieldType))
+                    {
+                        fieldType = typeof(Component);
+                    }
+                }
+            }
+
+            if (targetProperty.objectReferenceValue != null)
             {
                 return ("", null);
             }
-
-            // GetComponentInChildrenAttribute getComponentInChildrenAttribute = (GetComponentInChildrenAttribute) saintsAttribute;
-            Type fieldType = info.FieldType;
 
             Type type = getComponentInParentsAttribute.CompType ?? fieldType;
 
@@ -88,9 +126,19 @@ namespace SaintsField.Editor.Drawers
                     break;
                 }
 
-                componentInParent = isGameObject
-                    ? curCheckingTrans
-                    : curCheckingTrans.GetComponent(type);
+                if (isGameObject)
+                {
+                    componentInParent = curCheckingTrans;
+                }
+                else
+                {
+                    componentInParent = interfaceType == null
+                        ? curCheckingTrans.GetComponent(type)
+                        : curCheckingTrans.GetComponents(type).FirstOrDefault(interfaceType.IsInstanceOfType);
+                }
+                // componentInParent = isGameObject
+                //     ? curCheckingTrans
+                //     : curCheckingTrans.GetComponent(type);
 
 #if SAINTSFIELD_DEBUG && SAINTSFIELD_DEBUG_DRAW_PROCESS_GET_COMPONENT_IN_PARENTS
                 Debug.Log($"Search parent {levelLimit}, curCheckingTrans={curCheckingTrans}, componentInParent={componentInParent}");
@@ -118,11 +166,13 @@ namespace SaintsField.Editor.Drawers
                 }
                 else
                 {
-                    result = componentInParent.GetComponent(fieldType);
+                    result = interfaceType == null
+                        ? componentInParent.GetComponent(fieldType)
+                        : componentInParent.GetComponents(fieldType).FirstOrDefault(interfaceType.IsInstanceOfType);
                 }
             }
 
-            property.objectReferenceValue = result;
+            targetProperty.objectReferenceValue = result;
             return ("", result);
         }
 
@@ -141,7 +191,7 @@ namespace SaintsField.Editor.Drawers
 #if SAINTSFIELD_DEBUG && SAINTSFIELD_DEBUG_DRAW_PROCESS_GET_COMPONENT_IN_PARENTS
             Debug.Log($"GetComponentInParents DrawPostFieldUIToolkit for {property.propertyPath}");
 #endif
-            (string error, UnityEngine.Object result) = DoCheckComponent(property, (GetComponentInParentsAttribute)saintsAttribute, info);
+            (string error, UnityEngine.Object result) = DoCheckComponent(property, (GetComponentInParentsAttribute)saintsAttribute, info, parent);
             HelpBox helpBox = container.Q<HelpBox>(NamePlaceholder(property, index));
             if (error != helpBox.text)
             {