+++ Editor/Drawers/GetComponentInSceneAttributeDrawer.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Reflection;
 using SaintsField.Editor.Core;
 using SaintsField.Editor.Utils;
@@ -25,7 +27,7 @@ namespace SaintsField.Editor.Drawers
             int index,
             OnGUIPayload onGUIPayload, FieldInfo info, object parent)
         {
-            (string error, UnityEngine.Object result) = DoCheckComponent(property, saintsAttribute, info);
+            (string error, UnityEngine.Object result) = DoCheckComponent(property, saintsAttribute, info, parent);
             if (error != "")
             {
                 _error = error;
@@ -47,15 +49,53 @@ namespace SaintsField.Editor.Drawers
             ISaintsAttribute saintsAttribute, FieldInfo info, object parent) => _error == ""? position: ImGuiHelpBox.Draw(position, _error, EMessageType.Error);
         #endregion
 
-        private static (string error, UnityEngine.Object result) DoCheckComponent(SerializedProperty property, ISaintsAttribute saintsAttribute, FieldInfo info)
+        private static (string error, UnityEngine.Object result) DoCheckComponent(SerializedProperty property, ISaintsAttribute saintsAttribute, FieldInfo info, object parent)
         {
-            if (property.objectReferenceValue != null)
+            SerializedProperty targetProperty = property;
+            Type fieldType = info.FieldType;
+            Type interfaceType = null;
+            if (property.propertyType == SerializedPropertyType.Generic)
+            {
+                object propertyValue = SerializedUtils.GetValue(property, info, parent);
+
+                if (propertyValue is IWrapProp wrapProp)
+                {
+                    Type mostBaseType = SaintsInterfaceDrawer.GetMostBaseType(wrapProp.GetType());
+                    if (mostBaseType.IsGenericType && mostBaseType.GetGenericTypeDefinition() == typeof(SaintsInterface<,>))
+                    {
+                        IReadOnlyList<Type> genericArguments = mostBaseType.GetGenericArguments();
+                        if (genericArguments.Count == 2)
+                        {
+                            interfaceType = genericArguments[1];
+                        }
+                    }
+                    targetProperty = property.FindPropertyRelative(wrapProp.EditorPropertyName) ??
+                                     SerializedUtils.FindPropertyByAutoPropertyName(property,
+                                         wrapProp.EditorPropertyName);
+
+                    if(targetProperty == null)
+                    {
+                        return ($"{wrapProp.EditorPropertyName} not found in {property.propertyPath}", null);
+                    }
+
+                    SerializedUtils.FieldOrProp wrapFieldOrProp = Util.GetWrapProp(wrapProp);
+                    fieldType = wrapFieldOrProp.IsField
+                        ? wrapFieldOrProp.FieldInfo.FieldType
+                        : wrapFieldOrProp.PropertyInfo.PropertyType;
+
+                    if (interfaceType != null && fieldType != typeof(Component) && !fieldType.IsSubclassOf(typeof(Component)) && typeof(Component).IsSubclassOf(fieldType))
+                    {
+                        fieldType = typeof(Component);
+                    }
+                }
+            }
+
+            if (targetProperty.objectReferenceValue != null)
             {
                 return ("", null);
             }
 
             GetComponentInSceneAttribute getComponentInSceneAttribute = (GetComponentInSceneAttribute) saintsAttribute;
-            Type fieldType = info.FieldType;
 
             if (getComponentInSceneAttribute.CompType == typeof(GameObject))
             {
@@ -75,14 +115,19 @@ namespace SaintsField.Editor.Drawers
                     continue;
                 }
 
-                Component findSelfComponent = rootGameObject.GetComponent(type);
+                Component findSelfComponent = interfaceType == null
+                    ? rootGameObject.GetComponent(type)
+                    : rootGameObject.GetComponents(type).FirstOrDefault(interfaceType.IsInstanceOfType);
                 if (findSelfComponent != null)
                 {
                     componentInScene = findSelfComponent;
                     break;
                 }
 
-                Component findComponent = rootGameObject.GetComponentInChildren(type, includeInactive);
+                Component findComponent  = interfaceType == null
+                    ? rootGameObject.GetComponentInChildren(type, includeInactive)
+                    : rootGameObject.GetComponentsInChildren(type, includeInactive).FirstOrDefault(interfaceType.IsInstanceOfType);
+
                 // ReSharper disable once InvertIf
                 if (findComponent != null)
                 {
@@ -110,9 +155,9 @@ namespace SaintsField.Editor.Drawers
                 }
             }
 
-            if(!ReferenceEquals(property.objectReferenceValue, result))
+            if(!ReferenceEquals(targetProperty.objectReferenceValue, result))
             {
-                property.objectReferenceValue = result;
+                targetProperty.objectReferenceValue = result;
             }
             return ("", result);
         }
@@ -130,7 +175,7 @@ namespace SaintsField.Editor.Drawers
 #if SAINTSFIELD_DEBUG && SAINTSFIELD_DEBUG_DRAW_PROCESS_GET_COMPONENT_IN_SCENE
             Debug.Log($"GetComponent DrawPostFieldUIToolkit for {property.propertyPath}");
 #endif
-            (string error, UnityEngine.Object result) = DoCheckComponent(property, saintsAttribute, info);
+            (string error, UnityEngine.Object result) = DoCheckComponent(property, saintsAttribute, info, parent);
             HelpBox helpBox = container.Q<HelpBox>(NameHelpBox(property, index));
             if (error != helpBox.text)
             {